smyrna: remove unnecessary casts of 'ViewInfo.{w|h}'

+++ cmd/smyrna/glutrender.c
@@ -55,7 +55,7 @@ static void cb_reshape(int width, int height)
     /* setup various opengl things that we need */
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
-    init_arcBall(view->arcball, (float)view->w, (float)view->h);
+    init_arcBall(view->arcball, view->w, view->h);
     if (view->w > view->h) {
 	aspect = (float) view->w / (float) view->h;
 	glOrtho(-aspect * GL_VIEWPORT_FACTOR, aspect * GL_VIEWPORT_FACTOR,
